// convex/_generated/server.d.ts
type Env = {
  readonly CONVEX_CLOUD_URL: string;
  readonly CONVEX_SITE_URL: string;
  readonly AGENTMAIL_API_KEY: string | undefined;
  readonly AGENTMAIL_BASE_URL: string | undefined;
  readonly AGENTMAIL_INBOX_ID: string | undefined;
  readonly AGENTMAIL_WEBHOOK_SECRET: string | undefined;
  readonly FIRECRAWL_API_KEY: string;
  readonly FIRECRAWL_WEBHOOK_SECRET: string | undefined;
  readonly OPENAI_API_KEY: string;
};

/**
 * Define a query in this Convex app's public API.
 *
 * This function will be allowed to read your Convex database and will be accessible from the client.
 *
 * @param func - The query function. It receives a {@link QueryCtx} as its first argument.
 * @returns The wrapped query. Include this as an `export` to name it and make it accessible.
 */
export declare const query: QueryBuilder<DataModel, "public">;
// ...
export declare const internalQuery: QueryBuilder<DataModel, "internal">;

/**
 * Define a mutation in this Convex app's public API.
 *
 * This function will be allowed to modify your Convex database and will be accessible from the client.
 *
 * @param func - The mutation function. It receives a {@link MutationCtx} as its first argument.
 * @returns The wrapped mutation. Include this as an `export` to name it and make it accessible.
 */
export declare const mutation: MutationBuilder<DataModel, "public">;
// ...
export declare const internalMutation: MutationBuilder<DataModel, "internal">;
// ...
export declare const internalAction: ActionBuilder<DataModel, "internal">;
// ...
export declare const env: Env;